* Refactor DirectWrite into it's own folder. * Implement rendering bitmap and SVG by glyph indice using custom renderer. * Separate WIC and WIC's lib into two modules. * Fix not Releasing certain D2D objects.
