+++ pyglet/font/__init__.py
@@ -42,7 +42,7 @@ def _get_system_font_class() -> type[Font]:
     elif pyglet.compat_platform in ("win32", "cygwin"):
         from pyglet.libs.win32.constants import WINDOWS_7_OR_GREATER
         if WINDOWS_7_OR_GREATER and not pyglet.options["win32_gdi_font"]:
-            from pyglet.font.directwrite import Win32DirectWriteFont
+            from pyglet.font.dwrite import Win32DirectWriteFont
             _font_class = Win32DirectWriteFont
         else:
             from pyglet.font.win32 import GDIPlusFont
